+++ utils.js
@@ -1,0 +1,5 @@
+const isFunction = (fn) => typeof fn === 'function';
+
+export {
+    isFunction
+}